Adds tests for KeyCache and optimize former versioned group key loading
